requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
motor==3.3.2
pymongo==4.6.0
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
passlib==1.7.4
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
beanie==1.24.0
websockets==12.0
networkx==3.2.1
numpy==1.26.4
geopy==2.4.1
redis==5.0.1
aiofiles==23.2.1
groq==0.9.0
httpx==0.25.2